+++ src/main/java/com/webank/webase/node/mgr/base/tools/pagetools/List2Page.java
@@ -13,34 +13,28 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.webank.webase.node.mgr.base.tools.page;
+package com.webank.webase.node.mgr.base.tools.pagetools;
 
 import java.util.ArrayList;
-import java.util.Iterator;
 import java.util.List;
-import java.util.Map;
 
-public class Map2PagedList<T> {
-    private List<MapHandle> data;
+/**
+ * use PageData to transfer list to paged list
+ * @param <T>
+ */
+public class List2Page<T> {
+    private List<T> data;
     private Integer pageSize;
     private Integer pageIndex;
 
-    // map constructor
-    public Map2PagedList(Map<T,T> maps, Integer pageSize, Integer pageIndex) {
-        List<MapHandle> mapList = new ArrayList<>();
-        Iterator it = maps.keySet().iterator();
-        while (it.hasNext()) {
-            String key = it.next().toString();
-            MapHandle handle = new MapHandle(key, maps.get(key));
-            mapList.add(handle);
-        }
-        this.data = mapList;
+    public List2Page(List<T> data, Integer pageSize, Integer pageIndex) {
+        this.data = data;
         this.pageSize = pageSize;
         this.pageIndex = pageIndex;
     }
 
-    public List<MapHandle> getPagedList() {
-        List<MapHandle> resList = null; // 结果记录列表
+    public PageData<T> getResult() {
+        List<T> resList = null; // 结果记录列表
         Integer size = data.size(); // 总记录数
         Integer pages = size / pageSize; // 总页数
         if (size - pages * pageSize > 0)
@@ -50,8 +44,23 @@ public class Map2PagedList<T> {
         else if (pageIndex == pages)
             resList = data.subList((pageIndex - 1) * pageSize, size);
         else
-            resList = new ArrayList<>();
-        PageData<MapHandle> list = new PageData<>(resList, pageIndex, pageSize, resList.size(), size, pages);
+            resList = new ArrayList<T>();
+        return new PageData<T>(resList, pageIndex, pageSize, resList.size(), size, pages);
+    }
+
+    public List<T> getPagedList() {
+        List<T> resList = null; // 结果记录列表
+        Integer size = data.size(); // 总记录数
+        Integer pages = size / pageSize; // 总页数
+        if (size - pages * pageSize > 0)
+            ++pages;
+        if (pageIndex < pages)
+            resList = data.subList((pageIndex - 1) * pageSize, pageSize * pageIndex);
+        else if (pageIndex == pages)
+            resList = data.subList((pageIndex - 1) * pageSize, size);
+        else
+            resList = new ArrayList<T>();
+        PageData<T> list = new PageData<T>(resList, pageIndex, pageSize, resList.size(), size, pages);
         return list.getList();
     }
 }