+++ Changelog.md
@@ -1,3 +1,27 @@
+v1.2.1 (2019-11-22)
+
+**Add**
+- 邮件服务接口
+- 邮件告警类型接口
+- 告警日志接口
+- 证书管理增加SDK证书
+
+**Fix**
+- 优化：sh脚本支持secp256k1
+- 优化：升级fastjson为1.2.60, gradle-wrapper为v6.0.1
+- 优化：precompiled节点列表接口增加离线的共识/观察节点
+
+**兼容性**
+- 支持FISCO-BCOS v2.0.0-rc1 版本
+- 支持FISCO-BCOS v2.0.0-rc2 版本
+- 支持FISCO-BCOS v2.0.0-rc3 版本
+- 支持FISCO-BCOS v2.0.0 版本
+- WeBASE-Front v1.2.1
+- WeBASE-Node-Manager v1.2.1
+
+详细了解,请阅读[**技术文档**](https://webasedoc.readthedocs.io/zh_CN/latest/)。
+
+
 v1.2.0 (2019-10-29)
 
 **Add**
@@ -9,13 +33,13 @@ v1.2.0 (2019-10-29)
 - 支持FISCO-BCOS v2.0.0-rc2 版本
 - 支持FISCO-BCOS v2.0.0-rc3 版本
 - 支持FISCO-BCOS v2.0.0 版本
-- WeBASE-Front V1.2.0
-- WeBASE-Node-Manager V1.2.0
+- WeBASE-Front v1.2.0
+- WeBASE-Node-Manager v1.2.0
 
 详细了解,请阅读[**技术文档**](https://webasedoc.readthedocs.io/zh_CN/latest/)。
 
 
-V1.1.0 (2019-09-9)
+v1.1.0 (2019-09-9)
 
 **Add**
 - 查询cns
@@ -35,14 +59,14 @@ V1.1.0 (2019-09-9)
 - 支持FISCO-BCOS v2.0.0-rc2 版本
 - 支持FISCO-BCOS v2.0.0-rc3 版本
 - 支持FISCO-BCOS v2.0.0 版本
-- WeBASE-Web V1.1.0
-- WeBASE-Front V1.1.0
+- WeBASE-Web v1.1.0
+- WeBASE-Front v1.1.0
 
 详细了解,请阅读[**技术文档**](https://webasedoc.readthedocs.io/zh_CN/latest/)。
 
 
 
-### V1.0.0
+### v1.0.0
 
 (2019-06-27)